Redefinido campo Float de algunos modelos

+++ models/OPU.js
@@ -27,11 +27,11 @@ const OPU = db.define('OPU', {
         allowNull: false,
     },
     totalBolivares: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.FLOAT(50, 2),
         allowNull: false,
     },
     totalDolares: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.FLOAT(50, 2),
         allowNull: false,
     },
 });
